Wrapper hver søknad-app med PersonRouting for å kunne gjøre ulike kall for å hente person basert på hvilken stønadstype det gjelder Trenger også å route til ulike papirsøknader avhengig av stønadstypen
